消费者订单模块开发

+++ cloud-provider-payment8001/src/main/java/com/atguigu/springcloud/dao/PaymentDao.java
@@ -17,7 +17,7 @@ public interface PaymentDao {
      * @param payment 支付实体
      * @return id
      */
-    public Long create(Payment payment);
+    public int create(Payment payment);
 
     /**
      * 根据id查询
