+++ cloud-provider-payment8001/src/main/java/com/atguigu/springcloud/service/PaymentService.java
@@ -15,7 +15,7 @@ public interface PaymentService {
      * @param payment 支付实体
      * @return id
      */
-    public Long create(Payment payment);
+    public int create(Payment payment);
 
     /**
      * 根据id查询
