+++ cloud-provider-payment8001/src/main/java/com/atguigu/springcloud/service/impl/PaymentServiceImpl.java
@@ -20,7 +20,7 @@ public class PaymentServiceImpl implements PaymentService {
     private PaymentDao paymentDao;
 
     @Override
-    public Long create(Payment payment) {
+    public int create(Payment payment) {
         return paymentDao.create(payment);
     }
